Don't crash on non-atom alarm IDs
This fixes an issue where the Alarmist's alarm handler gets
automatically unregistered when an Alarm ID is sent that it's not
expecting. Normalization caught the issue, but the code ended up
returning a value that `:gen_event` wasn't expecting so it crashed.

This also makes normalization more explicit in its return values to be
more idiomatic. This allows the `:error` alarm ID (super poor name, but
seems like a silly special case.)

Unit tests were added to cover a trivial set/clear case and then to
exercise the unsupported alarm ID condition.