Make alarm handler more flexible when debugging
Normally the alarm handler should be swapped on init. That's what's
currently done. However when debugging, it's convenient to change out
the handler or recover it after a crash. That turns out to be a pain
since `:alarm_handler` is only the initial handler on SASL startup. This
makes the init code handle the other ways it could be added in a
reasonable way.