Replace Alarmist.current_alarms/0 with get_alarm/0
BACKWARDS INCOMPATIBLE CHANGE

This fixes a semantic issue where the function name indicated that
alarms were being return even though it was returning alarm IDs. It also
makes Alarmist consistent with the naming of the same function in
`:alarm_handler`.

Getting the list of alarm IDs is still useful so this also adds
`Alarmist.current_alarm_ids/0`.

This is a compile-time breaking change to force Alarmist users to force
users to look at the Alarmist version update and hope that they'll see
the `Alarmist.Event` breaking change as well.