Switch from PropertyTable events to a built-in one
BACKWARDS INCOMPATIBLE CHANGE

This fixes the TODO to create a custom event type for Alarmist. The main
advantage here is to make Alarmist easier to use since the PropertyTable
internals aren't exposed.  The second reason is that the PropertyTable
schema is changing and if we're going to break the API, we might as well
just do it once.