BREAKING: Rename defalarm to alarm_if
This clarifies that the expression following it is the condition that Alarmist
should use to set or clear the alarm.

In the future, there will be more kinds of ways to set and clear alarms.
`defalarm` made it sound like it should support those ways too, but that's
going to be challenging to make an all-in-one. At least, I'm not having any
luck with that strategy and `alarm_if` is sounding pretty good regardless.