Add alarm severities
This makes it possible to filter and sort through active alarms and also
to more easily ignore all of the temporary ones. Alarms severities
are the same values as Logger levels. The default is `:warning`.

This drops support for Elixir 1.15 to avoid handling the differences
with the pre-1.16 Logger.

This addresses #22 by making internal alarm IDs have `:debug` severity
and then making the default when getting alarms be to only include
`:info` and higher.