Fix startup race due to async handler initialization
This PR addresses a problem that probably affects unit tests the most
where there's a race between application start and a call to the
handler. Amazingly, this was going unnoticed in tests due to the return
value from `:gen_event.call` not being checked. Most functions only
return `:ok` based on spec, so this wasn't expected. When this
manifested in an error, it was quite a bit later when some state wasn't
available.

There are two main pieces to this PR. One is to add retries to the
`:gen_event.call` calls so that users don't have to worry about the
race. The calls still obey the 5 second timeout with the retries. In the
future, it might be desirable to expose the timeout to users like is
frequently done with `GenServer.call`.

The second part of the PR is to clean up Application start and stop.
This is needed for the unit tests to actually use the new code and
reproduce the race. Otherwise the handler never gets removed and the
swap silently fails.