Ensure intended execption is raised when not started
The retry mechanism would result in 0 or negative timeouts when the
handler is missing. This would cause `:gen_event.call/4` to raise when
the intention was for the nicer `RuntimeError` to raise. This change
makes it less confusing to debug since the negative timeouts really
looked like programming errors.

This also makes it possible to set the timeout on the one API call
that's being used to unit test this feature. This should be expanded to
the other API calls in the future.

Sadly, to be spec-compliant with `timeout/0`, `:infinity` needs to be
supported so that was added. Use at your own risk.