Add Alarmist.alarm_state/1 to get the state of 1 alarm
This makes it possible to get the state of one alarm quickly. Previously
the only way was to get all alarms and see if the one of interest was in
the the list.

As part of this change, `:unknown` was promoted as a state so that
callers could choose to be optimistic or pessimisitc on how to deal with
alarms that haven't been sent yet.