Fix unknown/clear ambiguities on alarms
This commit ensures event messages are always sent on alarm changes from
unknown -> clear and clear -> unknown. It also ensures that managed
alarms are always clear or set when they're actively being managed. This
results in more events being sent that previous, but this should be
welcome as it means that listeners will have more accurate state.

This fixes an issue where getting the state of a managed alarm could
return `:unknown` and the user would need to infer that it was `:clear`.
It was ok to do this since Alarmist internally did it. However, if you
did this, you couldn't tell if the managed alarm was never added or if
you mispelled the alarm ID. Now that managed alarms are always `:set` or
`:clear`, if you ever get `:unknown` back, you'll know that something is
wrong.